% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{AAclassify}
\alias{AAclassify}
\title{make Amino Acid classification}
\usage{
AAclassify(AAs)
}
\arguments{
\item{AAs}{A vector of amino acid in 1 letter code.}
}
\value{
Vector of classification with attribute 'encode' encoding proprieties
}
\description{
Classify AA according to biophysical properties
}
