% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{AAcolor}
\alias{AAcolor}
\title{Amino Acid Colors}
\usage{
AAcolor(AAs)
}
\arguments{
\item{AAs}{Amino Acids to take}
}
\value{
vector of HEX string representing the colors
}
\description{
Get the color that represent the vector of amino acids.
}
