% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{AAinteraction}
\alias{AAinteraction}
\title{Classify AA-AA interaction}
\usage{
AAinteraction(prop1s, prop2s)
}
\arguments{
\item{prop1s}{A vector of amino acid proprieties, atrribute 'encode' of AAclassify() output.}

\item{prop2s}{A vector of amino acid proprieties, atrribute 'encode' of AAclassify() output.}
}
\value{
Vector of interaction classification
}
\description{
Classify AA-AA interaction according to AA propaties.
}
