% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{aaPCboxPlot}
\alias{aaPCboxPlot}
\title{Amino acid type ploted against principal components in box plot}
\usage{
aaPCboxPlot(mono_motifs, Alignment, pos = "P-1", AApos = 13, PC = 1)
}
\arguments{
\item{mono_motifs}{List of motifs that are input as frequency values (0-1 with highest of 1)}

\item{Alignment}{The alignment resulted from concatAli}

\item{pos}{Binding matrix position to extract}

\item{AApos}{Residue position along the alignment}

\item{PC}{Principal componant to screen}
}
\value{
box plot
}
\description{
Make box plot of amino acid types and PC values
}
