% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{aaPCmap}
\alias{aaPCmap}
\title{Amino acid mapping against principal components}
\usage{
aaPCmap(mono_motifs, Alignment, pos = "P-1", AApos = 13, PC = 1, property = 1)
}
\arguments{
\item{mono_motifs}{List of motifs that are input as frequency values (0-1 with highest of 1)}

\item{Alignment}{The alignment resulted from concatAli}

\item{pos}{Binding matrix position to extract}

\item{AApos}{Residue position along the alignment}

\item{PC}{Principal componant to screen}

\item{property}{Biophysical property to screen. 1 = Hydropathy, 2 = Volume, 3 = Isoelectric Point}
}
\value{
Plot PC value against amino acid prooperty, returns summary of linear regression
}
\description{
Perform linear regression between the PC value of amino acid type at a given alignment position and
motif position and the selected amino acid property
}
