% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{AAsymbols}
\alias{AAsymbols}
\title{Amino Acid Symbols}
\usage{
AAsymbols(AAs)
}
\arguments{
\item{AAs}{Amino Acids to take}
}
\value{
vector of symbols for plotly
}
\description{
Get the symbol that represent the vector of amino acids.
}
