% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\docType{data}
\name{BLOSUM62}
\alias{BLOSUM62}
\title{BLOSUM62}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 25 rows and 25 columns.
}
\usage{
BLOSUM62
}
\description{
protein substitution matrix BLOSUM62 used in svd prediction
}
\references{
\url{https://web.mit.edu/~r/current/arch/i386_linux26/lib/R/library/Biostrings/html/substitution_matrices.html}
}
\author{
H. Pagès and P. Aboyoun
}
\keyword{2008.}
\keyword{23,}
\keyword{BLOSUM62}
\keyword{Bioinformatics,}
\keyword{Feb}
\keyword{K.}
\keyword{Malde,}
\keyword{The}
\keyword{alignment,}
\keyword{effect}
\keyword{from}
\keyword{imported}
\keyword{matrix}
\keyword{of}
\keyword{on}
\keyword{quality}
\keyword{sequence}
