% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{closestSeqPred}
\alias{closestSeqPred}
\title{Closest sequence prediction}
\usage{
closestSeqPred(mono_motifs, Alignment, pos = "P-1")
}
\arguments{
\item{mono_motifs}{List of motifs that are input as frequency values (0-1 with highest of 1)}

\item{Alignment}{The alignment resulted from concatAli}

\item{pos}{Binding matrix position to extract}
}
\value{
Data frame with two columns containing the true and predicted values
}
\description{
Use the Closest sequence prediction method (Weirauch, 2014) to predict the binding matrix
}
