% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{concatAli}
\alias{concatAli}
\title{Concatenate Alignment table}
\usage{
concatAli(Ali, start = 1, length = nchar(Ali[1, 2]))
}
\arguments{
\item{Ali}{Data frame with 2 columns, column 1 has names as identifiers and column 2 has a string of aligned sequence}

\item{start}{Starting position of the DNA-binding domain}

\item{length}{length of the DNA-binding domain}
}
\value{
Data frame with 2 columns: name and aligned sequence of DNA-binding domain
}
\description{
Concatenate .ali/.sto files from clustal omega and extract the alignment of the DNA-binding domain
}
