% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{CSpredPSAM}
\alias{CSpredPSAM}
\title{make predictions with Closest sequence}
\usage{
CSpredPSAM(predSeq, mono_motifs, Alignment)
}
\arguments{
\item{predSeq}{sequences to predict, matching alignment of train_alignment}

\item{mono_motifs}{Motif samples to include in the model, result from loadMono_motifs() or filterMotifList().}

\item{Alignment}{The alignment resulted from concatAli, ideally after matchAliMotif().}
}
\value{
predTrue data frame
}
\description{
Predict motif model with Closest sequence
}
