% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{ddG2frequency}
\alias{ddG2frequency}
\title{ddG matrix to Frequency matrix}
\usage{
ddG2frequency(matrix)
}
\arguments{
\item{matrix}{Motif matrix to be changed}
}
\value{
Motif matrix in frequency measurements
}
\description{
Change motif matrix in ddG measurements to frequency measurements
}
