% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{filterMotifList}
\alias{filterMotifList}
\title{Filter Motif List}
\usage{
filterMotifList(index, motifList, length = 8, posNames = paste0("P", 1:length))
}
\arguments{
\item{index}{filtered result for scoreMotifList()}

\item{motifList}{List of motifs, results from loadFromIndex(), used in scoreMotifList()}

\item{length}{length of motif alignment seed}

\item{posNames}{name of each motif position}
}
\value{
List of filtered motif models (frequency measurements) according to index
}
\description{
Filter from all motifs list for motifs according to a index data frame.
}
