% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{form.dddGList}
\alias{form.dddGList}
\title{Form dddG list}
\usage{
form.dddGList(
  Alignment,
  mono_motifs,
  keyPos = c(13, 14, 5, 26, 8),
  posMotif = "P1",
  leaveOut = c()
)
}
\arguments{
\item{Alignment}{Data frame of Alignment from concatAli}

\item{mono_motifs}{List of motifs output from loadMono_motifs()}

\item{keyPos}{Key positions along the alignment to screen}

\item{posMotif}{Position in the matrix to screen}

\item{leaveOut}{Samples from the alignment to be left out}
}
\value{
A list containing dddG matrices of key positions
}
\description{
Create dddG matrix of selected key positions
}
