% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{frequency2ddG}
\alias{frequency2ddG}
\title{Frequency matrix to ddG matrix}
\usage{
frequency2ddG(matrix)
}
\arguments{
\item{matrix}{Motif matrix to be changed}
}
\value{
Motif matrix in ddG measurements
}
\description{
Change motif matrix in frequency measurements to ddG measurements
}
