% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{gene2pos}
\alias{gene2pos}
\title{Gene to position-specific matrix}
\usage{
gene2pos(motifs, pos = "P1", nrow = 4)
}
\arguments{
\item{motifs}{List of motifs with $matrix containing a motif matrix, result form loadMono_motif()}

\item{pos}{Binding matrix position to extract}

\item{nrow}{Number of rows in motif matrices}
}
\value{
A position-specific motif matrix at the given motif position.
}
\description{
Creating a matrix of position-specific motif matrix at a given motif position
}
