% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{get.dddG}
\alias{get.dddG}
\title{tetrahedron dddG}
\usage{
get.dddG(motif1, motif2)
}
\arguments{
\item{motif1}{motif model of mutant in frequency}

\item{motif2}{motif model of WT in frequency}
}
\value{
calculated dddG model
}
\description{
Calculate dddG between two motifs with tetrahedron transformation
}
