% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{getBaseLineAccuracy}
\alias{getBaseLineAccuracy}
\title{Get base-line Accuracy}
\usage{
getBaseLineAccuracy(mono_motifs, pos, randomSample = FALSE)
}
\arguments{
\item{mono_motifs}{List of motifs that are input as frequency values (0-1 with highest of 1)}

\item{pos}{Position in motif to extract}

\item{randomSample}{If true, randomly select from studies of the sample protein; if fasle, takes the first two studies}
}
\value{
Data frame with two columns with pair-wise ddG values that can be plotted
}
\description{
Get the base-line accuracy for motif scoring according to different experiments performed on the same sample protein
}
