% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{getMatrix}
\alias{getMatrix}
\title{Get matrix}
\usage{
getMatrix(motifs, name, exactMatch = T)
}
\arguments{
\item{motifs}{List of motifs, result of loadMono_motifs()}

\item{name}{Name of sample to retrieve}

\item{exactMatch}{If ture, the name has to be exactly matching the name in the motifs list;
if false, grep is used to search.}
}
\value{
Motif Matrices with samples with the given name
}
\description{
Get matrix according to name from a list of motifs
}
