% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{getPvalTable}
\alias{getPvalTable}
\title{Get p-value table}
\usage{
getPvalTable(
  mono_motifs,
  Alignment,
  pos_index = c("P-3", "P-2", "P-1", "P1", "P2", "P3")
)
}
\arguments{
\item{mono_motifs}{List of motifs that are input as frequency values (0-1 with highest of 1)}

\item{Alignment}{The alignment resulted from concatAli}

\item{pos_index}{Position names of the consensus sequence}
}
\value{
Data frame of MANOVA p-values
}
\description{
Generate the p-value table of MANOVA between each amino acid residue-motif position combination
}
