% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{groupedPCC}
\alias{groupedPCC}
\title{Grouped PCC}
\usage{
groupedPCC(x, y, member = length(x), mean = T)
}
\arguments{
\item{x}{Vector 1}

\item{y}{Vector 2}

\item{member}{number of member of each group}

\item{mean}{If true, returns the mean of all groups; if false, returns a vector of each R-square value}
}
\value{
Average R-square value
}
\description{
Find average PCC value for two vectors separated in to groups with multiple members
}
