% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{groupedR2}
\alias{groupedR2}
\title{Grouped R-square}
\usage{
groupedR2(x, y, member = 4)
}
\arguments{
\item{x}{Vector 1}

\item{y}{Vector 2}

\item{member}{number of member of each group}
}
\value{
Average R-square value
}
\description{
Find average R-square value for two vectors separated in to groups with multiple members
}
