% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{inspectFeatureSVD}
\alias{inspectFeatureSVD}
\title{Inspect feature SVD}
\usage{
inspectFeatureSVD(svd, Alignment)
}
\arguments{
\item{svd}{Output from matrixSVD()}

\item{Alignment}{The alignment resulted from concatAli, ideally after matchAliMotif().}
}
\value{
List of 3 matrices, each matrix records the p-value of linear regression
 between SVD PCs and one of the three AAfeatures.
}
\description{
Inspect the correlation between amino acid properties and SVD PC values
}
