% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{JSON2Matrix}
\alias{JSON2Matrix}
\title{JSON to Matrix}
\usage{
JSON2Matrix(JSON_text, expo = exp(1), mode = 1)
}
\arguments{
\item{JSON_text}{a string containing the raw JSON information by readLines()}

\item{expo}{Factor of exponential to be applied to the matrix, 0 if no exponential should be applied}

\item{mode}{Binding mode to be parsed}
}
\value{
Binding matrix
}
\description{
Parse JSON file containing binding motifs to a scoring matrix
}
