% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{loadFromIndex}
\alias{loadFromIndex}
\title{Load From Index}
\usage{
loadFromIndex(index, modelFile_Template)
}
\arguments{
\item{index}{data frame of model information, contain to columns: gene_symbol and study}

\item{modelFile_Template}{dir of all model files, with specific identifier replaced as: $modelFile$}
}
\value{
List of all motif models
}
\description{
Load motif models in json format from a index form
}
