% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{loadJSONmotif}
\alias{loadJSONmotif}
\title{load motif from json}
\usage{
loadJSONmotif(modelFile, gene = "sampleGene", study = "NA")
}
\arguments{
\item{modelFile}{json file with consensus motif}

\item{gene}{gene symbol of the motif}

\item{study}{study of the motif}
}
\value{
list of motifs in the json file forward and backward
}
\description{
Load mono motif from JSON
}
