% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{makeSVDModel}
\alias{makeSVDModel}
\title{make SVD model}
\usage{
makeSVDModel(mono_motifs, Alignment, Positions, Ftest_pVal = 0.001)
}
\arguments{
\item{mono_motifs}{Motif samples to incldue in the model, result from loadMono_motifs() or filterMotifList().}

\item{Alignment}{The alignment resulted from concatAli, ideally after matchAliMotif().}

\item{Positions}{positions to train the model for}

\item{Ftest_pVal}{threshold used for F-test}
}
\value{
SVD regression model for entire motif
}
\description{
Make prediction of entire motif
}
