% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{makeSVDModel.singleMutation}
\alias{makeSVDModel.singleMutation}
\title{make SVD model.single mutation}
\usage{
makeSVDModel.singleMutation(mono_motifs, Alignment, Positions, mutation)
}
\arguments{
\item{mono_motifs}{Motif samples to incldue in the model, result from loadMono_motifs() or filterMotifList().}

\item{Alignment}{The alignment resulted from concatAli, ideally after matchAliMotif().}

\item{Positions}{positions to train the model for}

\item{mutation}{positions in the alignment where the mutation is seen}
}
\value{
SVD regression model for entire motif
}
\description{
Make prediction of entire motif for single mutation
}
