% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{matrix2seq}
\alias{matrix2seq}
\title{Matrix to Sequence}
\usage{
matrix2seq(scoringMatrix)
}
\arguments{
\item{scoringMatrix}{motif matrix with row names labeled as base types}
}
\value{
A string that contains the most preferred sequence of the scoring matrix
}
\description{
Output most preferred sequence of a motif matrix
}
