% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{matrix2tetrahedron}
\alias{matrix2tetrahedron}
\title{Matrix to Tetrahedron}
\usage{
matrix2tetrahedron(matrix)
}
\arguments{
\item{matrix}{A frequency matrix with 4 rows corresponding to the 4 bases ACGT}
}
\value{
A tetrahedron matrix with 3 columns representing the xyz coordinates
}
\description{
Change a frequency matrix into tetrahedron matrix
}
