% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{matrixReverse}
\alias{matrixReverse}
\title{Matrix reverse}
\usage{
matrixReverse(matrix)
}
\arguments{
\item{matrix}{Binding matrix in either energy, frequency, or information bit.}

\item{type}{One of 'energy', 'info', prob'}

\item{axes}{To show axes or not}

\item{motifPos}{To show reverse strand or not}

\item{labels}{To show labels or not}
}
\value{
A binding motif logo
}
\description{
Get the binding matrix of the reverse of a binding
}
