% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{matrixSVD}
\alias{matrixSVD}
\title{Matrix Singular Value Decomposition}
\usage{
matrixSVD(posMatrix)
}
\arguments{
\item{posMatrix}{Matrix of binding motifs in frequency measurements at a specific motif position, result of gene2pos()}
}
\value{
u matrix, v matrix, d vector from SVD, and the mean of each column of the transformed
tetrahedron coordinate matrix
}
\description{
Compute the tetrahedron SVD matrices of a position specific motif matrix of multiple samples
}
