% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{mononucleotide_logo}
\alias{mononucleotide_logo}
\title{Mono-nucleotide logo}
\usage{
mononucleotide_logo(
  matrix,
  type = "energy",
  axes = TRUE,
  reverse = FALSE,
  labels = TRUE
)
}
\arguments{
\item{matrix}{Binding matrix in either energy, frequency, or information bit.}

\item{type}{One of 'energy', 'info', prob'}

\item{axes}{To show axes or not}

\item{labels}{To show labels or not}

\item{motifPos}{To show reverse strand or not}
}
\value{
A binding motif logo
}
\description{
Create the logo of a mono-nucleotide binding matrix
}
