% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{normalize_sum1}
\alias{normalize_sum1}
\title{Normalize Sum to 1}
\usage{
normalize_sum1(matrix, rows = TRUE)
}
\arguments{
\item{matrix}{Matrix to be normalized}

\item{rows}{Logical that if true normalizes rows to 1, if false normalizes columns to 1}
}
\value{
Normalized matrix
}
\description{
Normalize the sum of each row or column of a matrix to 1
}
