% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{plot_tetrahedron_legend}
\alias{plot_tetrahedron_legend}
\title{Create Legend for Tetrahedron Plot}
\usage{
plot_tetrahedron_legend(
  sample_names,
  colors,
  symbols,
  point_size = 10,
  text_size = 12
)
}
\arguments{
\item{sample_names}{Character vector of sample names (should match colnames of posMatrix)}

\item{colors}{Vector of colors for each sample}

\item{symbols}{Vector of plotly symbols for each sample}

\item{point_size}{Size of legend symbols (default = 10)}

\item{text_size}{Font size for sample labels (default = 12)}
}
\value{
Plotly legend object
}
\description{
Generates a legend showing symbol and color for each sample in the tetrahedron plot
}
