% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{plot_tetrahedron_new}
\alias{plot_tetrahedron_new}
\title{Plot tetrahedron with customizable sample points}
\usage{
plot_tetrahedron_new(
  posMatrix,
  base_colors = c("green", "blue", "orange", "red"),
  size = 5,
  axis = FALSE,
  label = F,
  color_vector = NULL,
  vertex = F,
  alpha_vector = NULL,
  symbol_vector = NULL,
  size_vector = NULL
)
}
\arguments{
\item{posMatrix}{Matrix of binding motifs in frequency measurements at a specific motif position, result of gene2pos()}

\item{base_colors}{Color of bases at the vertexes}

\item{size}{Size of base dots in the tetrahedron (vertex points)}

\item{axis}{True to show axis}

\item{label}{Show label of data points according to colnames of posMatrix}

\item{vertex}{Show label of vertex}

\item{alpha_vector}{Optional vector of alpha values (0-1) for sample points}

\item{symbol_vector}{Optional vector of plotly symbol names for sample points}

\item{size_vector}{Optional vector of sizes for sample points}

\item{color}{Color sample points according to AA identity, only when colnames of posMatrix are 1 letter AA identifiers}
}
\value{
3D Plotly plot
}
\description{
Generate a tetrahedron representation system and plot all data samples at a given motif position
}
