% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{predict.dddGList}
\alias{predict.dddGList}
\title{Predict dddGList}
\usage{
\method{predict}{dddGList}(
  object,
  referenceMatrix,
  referenceSequence,
  targetSequence,
  keyPos = c(13, 14, 5, 8),
  posMotif = "P-1",
  ...
)
}
\arguments{
\item{object}{dddGList model output from form.dddGList()}

\item{referenceMatrix}{Binding matrix of the reference sample in ddG measurements}

\item{referenceSequence}{Protein sequence along the alignment of the reference sample}

\item{targetSequence}{Protien sequence along the alignmetn of the mutant sample}

\item{keyPos}{Key positions along the alignment to screen, must be included the dddGList}

\item{posMotif}{Position in the matrix to screen, must be included in the dddGList}

\item{...}{Place holder for generic function}
}
\value{
Binding matrix of the mutant sample at the selected matrix position
}
\description{
Predict energy given sequence and reference energy matrix and dddGList
}
