% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{predTrue.ddG2frequency}
\alias{predTrue.ddG2frequency}
\title{predTrue ddG to frequency}
\usage{
predTrue.ddG2frequency(predTrue, PFM = F)
}
\arguments{
\item{predTrue}{predTrue table, output of SVDregression.Iterative.CV()}

\item{PFM}{If true, will output PFM (column sum = 1) instead of PSAM(column max = 1)}
}
\value{
predTrue table
}
\description{
Change output of SVDregression.Iterative.CV() from ddG measurements to frequency measurements
}
