% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{pymolOpenFiles}
\alias{pymolOpenFiles}
\title{Pymol: open files}
\usage{
pymolOpenFiles(RA.files, pymol.dir, pml = "~/pymolBash.pml")
}
\arguments{
\item{RA.files}{vector of strings containing the path to the structure files}

\item{pymol.dir}{Path to pymol executable}

\item{pml}{Path to safe the pml file}
}
\value{
pymol program with structures loaded
}
\description{
Open structure files in pymol showing protiens in sticks and DNA in lines (.pdb/.cif)
}
