% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{RMSD}
\alias{RMSD}
\title{Root Mean Squared Difference}
\usage{
RMSD(a, b)
}
\arguments{
\item{a}{Vector 1}

\item{b}{Vector 2}
}
\value{
RMSD of the 2 vectors
}
\description{
Compute the RMSD between two vectors
}
