% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{run.pymol}
\alias{run.pymol}
\title{Run pymol}
\usage{
run.pymol(pymol.dir = "~/pyMOLWin.exe", pml = "~/pymolScript.pml", script)
}
\arguments{
\item{pymol.dir}{Path to pymol executable}

\item{pml}{Path to pml file}

\item{script}{Vector of String with each element containing a line of command}
}
\value{
promt: pymol started
}
\description{
Run pymol commands from R
}
