% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{Scorekmer}
\alias{Scorekmer}
\title{Score kmer}
\usage{
Scorekmer(
  kmers,
  motif,
  lflank = "",
  rflank = "",
  maxZero = T,
  additive = F,
  reverseComplement = T
)
}
\arguments{
\item{kmers}{List of kmers}

\item{motif}{motif model}

\item{lflank}{left flank added to the probe}

\item{rflank}{right flank added to the probe}

\item{maxZero}{rescale ddG values to maximize at 0 or center at 0}

\item{additive}{frequency or minimum energy}

\item{reverseComplement}{Score both forward and reverse complement}
}
\value{
Energy measurement for kmers
}
\description{
Score kmer probes with motif model
}
