% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{scoreMotifList}
\alias{scoreMotifList}
\title{Score Motif List}
\usage{
scoreMotifList(motifList, seed, weight = c())
}
\arguments{
\item{motifList}{List of motifs, results from loadFromIndex(), list of lists, each elements needs to have 'gene_symbol', 'study', 'mode', 'matrix'}

\item{seed}{seed matrix describing binding motif.}

\item{weight}{numeric vector of weights of each motif position of the seed}
}
\value{
List of all motif models
}
\description{
Score a List of motifs
}
