% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{scoreMotifSeed}
\alias{scoreMotifSeed}
\title{Score Motif Seed}
\usage{
scoreMotifSeed(matrix, seed, weight = c())
}
\arguments{
\item{matrix}{motif matrix in frequency measurements}

\item{seed}{seed matrix describing binding motif.}

\item{weight}{numeric vector of weights of each motif position of the seed}
}
\value{
Minimum distance score and position
}
\description{
Score a motif matrix with a seed to find the position of a motif
}
