% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{SRdist}
\alias{SRdist}
\title{Similarity Regression Distance}
\usage{
SRdist(seq1, seq2, weight)
}
\arguments{
\item{seq1}{Sequence 1}

\item{seq2}{Sequence 2}

\item{weight}{pretrained weights for SR}
}
\value{
sequence distance
}
\description{
Calculate distance of similarity regression
}
