% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{SRpredPSAM}
\alias{SRpredPSAM}
\title{make predictions with Similarity Regression}
\usage{
SRpredPSAM(predSeq, mono_motifs, Alignment, weightfile)
}
\arguments{
\item{predSeq}{sequences to predict, matching alignment of train_alignment}

\item{mono_motifs}{Motif samples to incldue in the model, result from loadMono_motifs() or filterMotifList().}

\item{Alignment}{The alignment resulted from concatAli, ideally after matchAliMotif().}

\item{weightfile}{json file that contains pretrained weights from Lamber et. al.,}
}
\value{
predTrue data frame
}
\description{
Predict motif model with Similarity regression
}
