% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{svdANOVA}
\alias{svdANOVA}
\title{SVD ANOVA on residue types}
\usage{
svdANOVA(svd, Alignment)
}
\arguments{
\item{svd}{Result from matrixSVD()}

\item{Alignment}{Alignment table with name and aligned sequences in the same order as the matrixSVD input}
}
\value{
A data frame of -log10 p-values with cols = number of principal components in SVD, rows = number of positions in the alignment
}
\description{
Perform ANOVA test between the u-matrix and the residue types along the alignment
}
