% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{SVDmodelPredPSAM}
\alias{SVDmodelPredPSAM}
\title{Predict PSAM with SVD model}
\usage{
SVDmodelPredPSAM(SVDmodel, predSeq, zero = 0.01)
}
\arguments{
\item{SVDmodel}{pre-trained SVDmodel from makeSVDModel()}

\item{predSeq}{sequence to predict, matching alignment of train_alignment}

\item{zero}{Lower limit to be assigned to predicted values}

\item{pos}{positions to show in the PSAM}
}
\value{
A PSAM in frequency measurements
}
\description{
predict for multiple positions and output PSAM with pre-trained SVDmodel
}
