% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{SVDpredPSAM}
\alias{SVDpredPSAM}
\title{Predict PSAM}
\usage{
SVDpredPSAM(
  train_motifs,
  train_alignment,
  predSeq,
  pos = c("P-3", "P-2", "P-1", "P1", "P2", "P3"),
  zero = 0.001,
  useSimilarAA = T
)
}
\arguments{
\item{train_motifs}{motifs for training, result from loadMono_motifs()}

\item{train_alignment}{alignment for training, result for matchAliMotif()}

\item{predSeq}{sequences to predict, matching alignment of train_alignment}

\item{pos}{positions to show in the PSAM}

\item{zero}{Lower limit to be assigned to predicted values}

\item{useSimilarAA}{Use the weighted sum of similar AA when no matching AA is found in training data}
}
\value{
A PSAM in frequency measruements
}
\description{
predict for multiple positions and output PSAM
}
