% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{tetrahedron2matrix}
\alias{tetrahedron2matrix}
\title{Tetrahedron to Matrix}
\usage{
tetrahedron2matrix(tetraMatrix)
}
\arguments{
\item{tetraMatrix}{Tetrahedron coordinates, result from matrix2tetrahedron}
}
\value{
A PSSM
}
\description{
Transform tetrahedron coordinate representation of PSSM to PSSM
}
