% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{trainSVD.Iterative}
\alias{trainSVD.Iterative}
\title{Train SVD-regression model with Iterative method}
\usage{
trainSVD.Iterative(svd, Alignment, Ftest_pVal = 0.1)
}
\arguments{
\item{svd}{Result from matrixSVD()}

\item{Alignment}{Alignment table with name and aligned sequences in the same order as the matrixSVD input}

\item{Ftest_pVal}{Threshold for F-test}
}
\value{
SVD-regression model
}
\description{
Train a SVD-regression model with matrixSVD and matched Alignment with Iterative method
}
