% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{trainSVD}
\alias{trainSVD}
\title{Train SVD-regression model}
\usage{
trainSVD(svd, Alignment, no.keyPos = c(4, 3, 2), keyPos = c())
}
\arguments{
\item{svd}{Result from matrixSVD()}

\item{Alignment}{Alignment table with name and aligned sequences in the same order as the matrixSVD input}

\item{no.keyPos}{Number of key positions to use for each PC, length of this vector should be the same as
the length of the svd$d vector}

\item{keyPos}{A list of key positions to use for prediction, using this argument stops key position identification.
no.keyPos should still be entered for the number of keyPos used for each PC. If using 0 keyPos for a PC,
enter a valid keypos number and put 0 in the no.keyPos vector for this PC.}
}
\value{
SVD-regression model
}
\description{
Train a SVD-regression model with matrixSVD and matched Alignment
}
