% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{weightedMean}
\alias{weightedMean}
\title{Weighted mean}
\usage{
weightedMean(vector, weight)
}
\arguments{
\item{vector}{A vector of numbers}

\item{weight}{weight of each number, should have the same length as vector}
}
\value{
Weighted mean
}
\description{
Compute weighted mean of a vector of numbers according to a vector of weight
}
