% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{write.bash}
\alias{write.bash}
\title{Write Bash}
\usage{
write.bash(script, outfile)
}
\arguments{
\item{script}{Vector of string that are lines of UNIX commands}

\item{outfile}{Path and name of bash file to be output}
}
\value{
successful transformation shows a reminder text: "bash script done, remember to dos2unix"
}
\description{
Write bash file with UNIX commands
}
