% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{writeFasta}
\alias{writeFasta}
\title{Write Fasta}
\usage{
writeFasta(data, filename)
}
\arguments{
\item{data}{data.frame that has two columns: name and seq, the colnames is crucial}

\item{filename}{Path and name of fasta file to be output}
}
\value{
Nothing
}
\description{
Write fasta files from data frame containing name and sequence
}
