% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FamilyCode-package.R
\docType{package}
\name{FamilyCode-package}
\alias{FamilyCode}
\alias{FamilyCode-package}
\title{FamilyCode: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
