% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{loadMono_motifs}
\alias{loadMono_motifs}
\title{Load mono_motifs}
\usage{
loadMono_motifs(
  bHLH_index,
  modelFile_Template = "~/$modelFile$/result/fit.models.consensus.json",
  rec_seq = "CANNTG",
  pos_index = c("P-3", "P-2", "P-1", "P1", "P2", "P3"),
  checkSymmetry = FALSE,
  withTable = TRUE
)
}
\arguments{
\item{bHLH_index}{data frame containing the identifiers of each experiment,
contains protein name as $gene_symbol and study name as $study in the first two column,
ProBound running folders should be name as $gene_symbol_$study.}

\item{modelFile_Template}{Path to a sample fit.models.consensus.json and switching
the $gene_symbol_$study identifier to $modelFile$.}

\item{rec_seq}{Consensus sequence to be recognized, use N for variable base positions}

\item{pos_index}{Position names of the consensus sequence}

\item{checkSymmetry}{To check symmetry like bHLH protein, should be changed for families other than bHLH}

\item{withTable}{If true, output with bHLH_index table with scoring and model number; if false, only output
list mono_motifs.}
}
\value{
List of motifs in mono_motifs, also an info table if withTable == T
}
\description{
Load binding motifs from dir containing ProBound results, the model with the highest
consensus sequence recognition score will be added to the resulting list
}
