% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{matchAliMotif}
\alias{matchAliMotif}
\title{Match Alignment and Motif}
\usage{
matchAliMotif(mono_motifs, Alignment)
}
\arguments{
\item{mono_motifs}{List of motifs output from loadMono_motifs}

\item{Alignment}{Data frame of Alignment from concatAli}
}
\value{
A filtered and reordered Alignment table
}
\description{
Match the order of mono_motifs list and Alignment data frame
}
