% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{matrixReverse}
\alias{matrixReverse}
\title{Matrix reverse}
\usage{
matrixReverse(matrix)
}
\arguments{
\item{matrix}{Binding matrix in either ddG or frequency.}
}
\value{
The reversed binding motif matrix
}
\description{
Get the binding matrix of the reverse of a binding
}
