% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{plot_tetrahedron}
\alias{plot_tetrahedron}
\title{Plot tetrahedron}
\usage{
plot_tetrahedron(
  posMatrix,
  base_colors = c("green", "blue", "orange", "red"),
  size = 5,
  axis = FALSE,
  label = F,
  color = F,
  vertex = F
)
}
\arguments{
\item{posMatrix}{Matrix of binding motifs in frequency measurements at a specific motif position, result of gene2pos()}

\item{base_colors}{Color of bases at the vertexes}

\item{size}{Size of dots in the tetrahedron}

\item{axis}{True to show axis}

\item{label}{Show label of data points according to colnames of posMatrix}

\item{color}{Color sample points according to AA identity, only when colnames of posMatrix are 1 letter AA identifiers}

\item{vertex}{Show label of vertex}
}
\value{
3D Plotly plot
}
\description{
Generate a tetrahedron representation system and plot all data samples at a given motif position
}
