% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{predict.SVD}
\alias{predict.SVD}
\title{Predict with SVD-regression model}
\usage{
\method{predict}{SVD}(object, Alignment, zero = 0.001, useSimilarAA = F, ...)
}
\arguments{
\item{object}{SVD-regression model result from trainSVD()}

\item{Alignment}{Alignment table with name and aligned sequences to be predicted}

\item{zero}{Lower limit to be assigned to predicted values}

\item{useSimilarAA}{Use the weighted sum of similar AA when no matching AA is found in training data}

\item{...}{Place holder for generic function}
}
\value{
Matrix of binding motif models in frequency measurements. attr('confidence') give confidence estimates of prediction
}
\description{
Predict binding matrix with SVD-regression model
}
