% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCfunctions.R
\name{SVDregression.Iterative.CV}
\alias{SVDregression.Iterative.CV}
\title{SVD-regression Cross-validation for Iterative method}
\usage{
SVDregression.Iterative.CV(
  mono_motifs,
  Alignment,
  pos = "P-1",
  Ftest_pVal = 0.001,
  zero = 0.01,
  useSimilarAA = F
)
}
\arguments{
\item{mono_motifs}{Motif samples to incldue in the CV test, result from loadMono_motifs().}

\item{Alignment}{The alignment resulted from concatAli, ideally after matchAliMotif().}

\item{pos}{Position of the binding matrix to predict.}

\item{Ftest_pVal}{Threshold for F-test, only used when iterative is True.}

\item{zero}{Lower limit to be assigned to predicted values}

\item{useSimilarAA}{Use AA similarity assesment to fill unseen AAs}
}
\value{
A data frame with two columns of true and predicted values
}
\description{
Leave-one-out cross-validation for SVD-regression model with Iterative feature selection method
}
